New branch for validator development with unfinished changes
